+++ packages/xml-views-validation/test/unit/validators/attributes/invalid-boolean-value.test.ts
@@ -15,7 +15,7 @@ describe("the invalid boolean value validation", () => {
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.49",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);