+++ packages/xml-views-validation/test/unit/validators/document/non-unique-id.test.ts
@@ -27,7 +27,7 @@ describe("the use of non unique id validation", () => {
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.49",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);