+++ packages/xml-views-tooltip/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## 4.0.22
+
+### Patch Changes
+
+- a9a2141: Add logger package
+- Updated dependencies [a9a2141]
+  - @ui5-language-assistant/semantic-model@4.0.15
+  - @ui5-language-assistant/logic-utils@4.0.16
+  - @ui5-language-assistant/context@4.0.22
+
+## 4.0.21
+
+### Patch Changes
+
+- 910e437: Adaptation to the latest UI5 maintenance changes
+- Updated dependencies [910e437]
+  - @ui5-language-assistant/semantic-model@4.0.14
+  - @ui5-language-assistant/context@4.0.21
+  - @ui5-language-assistant/logic-utils@4.0.15
+
 ## 4.0.20
 
 ### Patch Changes